add: new time capsule page with dynamic routing and UUID generation
This commit adds a new Svelte component for the time capsule page. The component includes a script that imports the `onMount` function from the 'svelte' library, as well as other necessary dependencies. On mount, the script checks if a `capsule_id` parameter is present in the URL. If not, it generates a UUID using the `uuidv4` function and redirects to a new route with the generated ID. The component also sets the title of the page to "Capsule" using `<svelte:head>`.