feat: Add locked_minutes field to Asset type
This commit adds a new field `locked_minutes` to the `Asset` type in the `main.mo` file. The `locked_minutes` field represents the number of minutes an asset should remain locked.

It also introduces a new function `add_file_with_time` in the same file, which allows adding a file with a specified number of locked minutes.

In addition, this commit updates the UI code in the `+page.svelte` file to include an input field for setting the number of locked minutes when adding a file. The value entered in this input field is passed as an argument to the `add_file_with_time` function.

The CSS code in the `output.css` file has been modified to include styles for a new class `.w-1/2`, which sets the width of an element to 50%.